Move to a verison of the pystream which pins setuptools-scm
Currently deployments of this package into the compbox do so into
the global Python space, which uses the system-installed setuptools
which is older than the latest version of setuptools-scm supports.
Move to a version of pystream which pins an older version of
setuptools-scm so that such installs continue to work for the time
being.